Dropped Unsync'd Sar Functionality in favour of single full SAR
