+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsremandandsentencingapi/config/SubjectAccessRequestConfiguration.kt
@@ -1,6 +1,5 @@
 package uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.config
 
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.jpa.repository.subjectaccessrequest.CourtCaseSarRepository
@@ -8,44 +7,24 @@ import uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.jpa.repository.s
 import uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.jpa.repository.subjectaccessrequest.RecallSarRepository
 import uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.service.CourtRegisterService
 import uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.service.PersonService
-import uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.service.subjectaccessrequest.AllPrisonerDetailsService
 import uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.service.subjectaccessrequest.PrisonerDetailsService
-import uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.service.subjectaccessrequest.UnsyncedPrisonerDetailsService
 
 @Configuration
 @ConditionalOnSarEnabled
 class SubjectAccessRequestConfiguration {
 
-  @ConditionalOnProperty(
-    prefix = "hmpps.sar.mode.all-sar-data",
-    name = ["enabled"],
-    havingValue = "true",
-    matchIfMissing = false,
-  )
   @Bean
-  fun allPrisonerDetailsService(
+  fun prisonerDetailsService(
     courtCaseSarRepository: CourtCaseSarRepository,
     recallSarRepository: RecallSarRepository,
     immigrationDetentionSarRepository: ImmigrationDetentionSarRepository,
     personService: PersonService,
     courtRegisterService: CourtRegisterService,
-  ): PrisonerDetailsService = AllPrisonerDetailsService(
+  ): PrisonerDetailsService = PrisonerDetailsService(
     courtCaseSarRepository,
     recallSarRepository,
     immigrationDetentionSarRepository,
     personService,
     courtRegisterService,
   )
-
-  @ConditionalOnProperty(
-    prefix = "hmpps.sar.mode.all-sar-data",
-    name = ["enabled"],
-    havingValue = "false",
-    matchIfMissing = true,
-  )
-  @Bean
-  fun unsyncedPrisonerDetailsService(
-    immigrationDetentionUnsyncedSarRepository: ImmigrationDetentionSarRepository,
-    courtCaseUnsyncedSarRepository: CourtCaseSarRepository,
-  ): PrisonerDetailsService = UnsyncedPrisonerDetailsService(immigrationDetentionUnsyncedSarRepository, courtCaseUnsyncedSarRepository)
 }
