+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppsremandandsentencingapi/integration/subjectaccessrequest/util/ExpectResponseData.kt
@@ -2,14 +2,16 @@ package uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.integration.sub
 
 object ExpectResponseData {
 
-  fun emptyNotInNomisResponse(): String =
-    """
-         {
-          "attachments": [],
-          "content": {
-            "prisonerNumber": "PRI123",
-            "immigrationDetentions": []
-          }
+  fun emptyFullDataResponse(): String = """
+        {
+            "content": {
+                "prisonerNumber": "PRI123",
+                "prisonerName": "Cormac Meza",
+                "courtCases": [],
+                "recalls": [],
+                "immigrationDetentions": []
+            },
+            "attachments": []
         }
-    """.trimIndent()
+  """.trimIndent()
 }
