--- src/test/resources/application-test-all-sar-data.yml
@@ -1,11 +1,0 @@
-hmpps:
-  sar:
-    mode:
-      all-sar-data:
-        enabled: true
-    template:
-      path: /templates/template_remand-and-sentencing-all-data-api.mustache
-    tests:
-      expected-api-response:
-        path: /sar/alldata/sar-api-response.json
-      attachments-expected: false
