+++ src/test/resources/application-test-sar.yml
@@ -1,9 +1,6 @@
 hmpps:
   sar:
     enabled: true
-    mode:
-      all-sar-data:
-        enabled: false
     additionalAccessRole: ROLE_VIEW_PRISONER_DATA
     template:
       path: /templates/template_remand-and-sentencing-api.mustache
