+++ src/test/resources/sar/sar-api-response.json
@@ -1,15 +1,166 @@
 {
   "prisonerNumber": "PRI123",
-  "immigrationDetentions": [
+  "prisonerName": "Cormac Meza",
+  "courtCases": [
     {
-      "homeOfficeReferenceNumber": "3240593452",
-      "noLongerOfInterestReason": "RIGHT_TO_REMAIN",
-      "noLongerOfInterestComment": "Currently has the right to Remain"
+      "courtName": "Crown Court #1",
+      "caseStatus": "ACTIVE",
+      "createdAt": "${json-unit.regex}^\\d{4}-\\d{2}-\\d{2} \\d{2}:\\d{2}.*$",
+      "updatedAt": "${json-unit.regex}^\\d{4}-\\d{2}-\\d{2} \\d{2}:\\d{2}.*$",
+      "latestCourtAppearance": {
+        "appearanceDate": "${json-unit.regex}^\\d{4}-\\d{2}-\\d{2}$",
+        "appearanceOutcomeName": "Imprisonment",
+        "warrantType": "SENTENCING",
+        "convictionDate": "${json-unit.regex}^\\d{4}-\\d{2}-\\d{2}$",
+        "nextAppearanceDate": "${json-unit.regex}^\\d{4}-\\d{2}-\\d{2}$",
+        "criminalAppealOffenceReference": "No Data Held",
+        "charges": [
+          {
+            "offenceCode": "AA06027",
+            "offenceDescription": "Veterinary surgeon fail to notify incorrect certification",
+            "terrorRelated": "Yes",
+            "foreignPowerRelated": "No",
+            "domesticViolenceRelated": "No",
+            "offenceStartDate": "${json-unit.regex}^\\d{4}-\\d{2}-\\d{2}$",
+            "offenceEndDate": "No Data Held",
+            "chargeOutcome": "Imprisonment",
+            "liveSentence": {
+              "sentenceTypeDescription": "Serious Offence Sec 250 Sentencing Code (U18)",
+              "sentenceTypeClassification": "STANDARD",
+              "periodLengths": [
+                {
+                  "years": 1,
+                  "months": 0,
+                  "weeks": 0,
+                  "days": 0,
+                  "periodOrder": "years"
+                }
+              ],
+              "sentenceServeType": "FORTHWITH"
+            }
+          },
+          {
+            "offenceCode": "EC10001",
+            "offenceDescription": "Fail to comply /contravene provision in schedule 3 of Eggs and Chicks (Wales) Regulations 2010",
+            "terrorRelated": "Yes",
+            "foreignPowerRelated": "No",
+            "domesticViolenceRelated": "No",
+            "offenceStartDate": "${json-unit.regex}^\\d{4}-\\d{2}-\\d{2}$",
+            "offenceEndDate": "No Data Held",
+            "chargeOutcome": "Imprisonment",
+            "liveSentence": {
+              "sentenceTypeDescription": "Serious Offence Sec 250 Sentencing Code (U18)",
+              "sentenceTypeClassification": "STANDARD",
+              "periodLengths": [
+                {
+                  "years": 1,
+                  "months": 0,
+                  "weeks": 0,
+                  "days": 0,
+                  "periodOrder": "years"
+                }
+              ],
+              "sentenceServeType": "FORTHWITH"
+            }
+          }
+        ]
+      },
+      "appearances": [
+        {
+          "appearanceDate": "${json-unit.regex}^\\d{4}-\\d{2}-\\d{2}$",
+          "appearanceOutcomeName": "Imprisonment",
+          "warrantType": "SENTENCING",
+          "convictionDate": "${json-unit.regex}^\\d{4}-\\d{2}-\\d{2}$",
+          "nextAppearanceDate": "${json-unit.regex}^\\d{4}-\\d{2}-\\d{2}$",
+          "criminalAppealOffenceReference": "No Data Held",
+          "charges": [
+            {
+              "offenceCode": "AA06027",
+              "offenceDescription": "Veterinary surgeon fail to notify incorrect certification",
+              "terrorRelated": "Yes",
+              "foreignPowerRelated": "No",
+              "domesticViolenceRelated": "No",
+              "offenceStartDate": "${json-unit.regex}^\\d{4}-\\d{2}-\\d{2}$",
+              "offenceEndDate": "No Data Held",
+              "chargeOutcome": "Imprisonment",
+              "liveSentence": {
+                "sentenceTypeDescription": "Serious Offence Sec 250 Sentencing Code (U18)",
+                "sentenceTypeClassification": "STANDARD",
+                "periodLengths": [
+                  {
+                    "years": 1,
+                    "months": 0,
+                    "weeks": 0,
+                    "days": 0,
+                    "periodOrder": "years"
+                  }
+                ],
+                "sentenceServeType": "FORTHWITH"
+              }
+            },
+            {
+              "offenceCode": "EC10001",
+              "offenceDescription": "Fail to comply /contravene provision in schedule 3 of Eggs and Chicks (Wales) Regulations 2010",
+              "terrorRelated": "Yes",
+              "foreignPowerRelated": "No",
+              "domesticViolenceRelated": "No",
+              "offenceStartDate": "${json-unit.regex}^\\d{4}-\\d{2}-\\d{2}$",
+              "offenceEndDate": "No Data Held",
+              "chargeOutcome": "Imprisonment",
+              "liveSentence": {
+                "sentenceTypeDescription": "Serious Offence Sec 250 Sentencing Code (U18)",
+                "sentenceTypeClassification": "STANDARD",
+                "periodLengths": [
+                  {
+                    "years": 1,
+                    "months": 0,
+                    "weeks": 0,
+                    "days": 0,
+                    "periodOrder": "years"
+                  }
+                ],
+                "sentenceServeType": "FORTHWITH"
+              }
+            }
+          ]
+        }
+      ]
     },
     {
-      "homeOfficeReferenceNumber": "5340593452",
-      "noLongerOfInterestReason": "BRITISH_CITIZEN",
-      "noLongerOfInterestComment": "Recently made a British Citizen"
+      "courtName": "No Data Held",
+      "caseStatus": "ACTIVE",
+      "createdAt": "${json-unit.regex}^\\d{4}-\\d{2}-\\d{2} \\d{2}:\\d{2}.*$",
+      "updatedAt": "${json-unit.regex}^\\d{4}-\\d{2}-\\d{2} \\d{2}:\\d{2}.*$",
+      "latestCourtAppearance": null,
+      "appearances": [
+        {
+          "appearanceDate": "${json-unit.regex}^\\d{4}-\\d{2}-\\d{2}$",
+          "appearanceOutcomeName": "Immigration Detainee",
+          "warrantType": "IMMIGRATION",
+          "convictionDate": "No Data Held",
+          "nextAppearanceDate": "No Data Held",
+          "criminalAppealOffenceReference": "No Data Held",
+          "charges": []
+        }
+      ]
+    }
+  ],
+  "recalls": [
+    {
+      "recallType": "FTR_14",
+      "revocationDate": "${json-unit.regex}^\\d{4}-\\d{2}-\\d{2}$",
+      "returnToCustodyDate": "${json-unit.regex}^\\d{4}-\\d{2}-\\d{2}$",
+      "inPrisonOnRevocationDate": "No",
+      "recallSentenceStatus": "ACTIVE"
+    }
+  ],
+  "immigrationDetentions": [
+    {
+      "immigrationDetentionRecordType": "IS91",
+      "homeOfficeReferenceNumber": "No Data Held",
+      "recordDate": "${json-unit.regex}^\\d{4}-\\d{2}-\\d{2}$",
+      "noLongerOfInterestReason": "No Data Held",
+      "noLongerOfInterestComment": "No Data Held"
     }
   ]
 }
